+++ packages/radio/components/RadioGroup.tsx
@@ -1,0 +1,91 @@
+"use client";
+
+import * as React from "react";
+import { Radio, type IRadio } from "./Radio";
+import clsx from "clsx";
+
+export interface IRadioGroup {
+  name: string;
+  value: string;
+  children: React.ReactElement<IRadio> | React.ReactElement<IRadio>[];
+  onChange?: (newValue: string) => void;
+  disabled?: boolean;
+  size?: "xsm" | "sm" | "md" | "lg" | "xlg";
+  variant?:
+    | "default"
+    | "primary"
+    | "secondary"
+    | "success"
+    | "error"
+    | "warning";
+  appearance?:
+    | "solid"
+    | "outline"
+    | "outline-dashed"
+    | "no-outline"
+    | "glowing";
+  align?: "horizontal" | "vertical";
+  alignLabel?: "top" | "right" | "bottom" | "left";
+  highlightLabel?: boolean;
+  className?: string;
+  style?: React.CSSProperties;
+}
+
+export const RadioGroup: React.FC<IRadioGroup> = ({
+  name,
+  value,
+  onChange,
+  children,
+  disabled = false,
+  highlightLabel = false,
+  size = "md",
+  variant = "default",
+  appearance = "outline",
+  align = "horizontal",
+  alignLabel = "right",
+  className = "",
+  style = {},
+}) => {
+  const uniqueId = React.useId();
+  const [currentValue, setCurrentValue] = React.useState<string>(value);
+
+  const radios = React.Children.toArray(
+    children
+  ) as React.ReactElement<IRadio>[];
+
+  const handleChange = (currentRadioValue: string) => {
+    if (onChange) onChange(currentRadioValue);
+    setCurrentValue(currentRadioValue);
+  };
+
+  return (
+    <div
+      key={uniqueId}
+      className={clsx(
+        "radio-group",
+        `radio-group--${align}`,
+        `radio-group-size--${size}`,
+        `radio-group-variant--${variant}`,
+        `radio-group-appearance--${appearance}`,
+        className
+      )}
+      style={style}
+    >
+      {radios.map(({ props }) => (
+        <Radio
+          {...props}
+          key={`${name}-${props.value}-${uniqueId}`}
+          name={`${name}-${props.value}-${uniqueId}`}
+          checked={props.value === currentValue}
+          onChange={() => handleChange(props.value)}
+          disabled={props.disabled ?? disabled}
+          size={props.size ?? size}
+          variant={props.variant ?? variant}
+          appearance={props.appearance ?? appearance}
+          highlightLabel={props.highlightLabel ?? highlightLabel}
+          alignLabel={props.alignLabel ?? alignLabel}
+        />
+      ))}
+    </div>
+  );
+};