+++ packages/utils/helpers/index.ts
@@ -1,6 +1,2 @@
 export * from "./ripple";
 export * from "./Portal";
-export * from "./mapKeys";
-export * from "./kebabCase";
-export * from "./omit";
-export * from "./clsx";