+++ packages/utils/helpers/kebabCase.ts
@@ -1,3 +1,0 @@
-export const kebabCase = (s: string) => {
-  return s.replace(/([a-z])([A-Z])/g, "$1-$2").toLowerCase();
-};